Update types for (in)equality checks
For equalities and inequalities, we can change the types on each side to
always convert to the larger type, which ensures all conversions are
safe.

+++ src/vmm/src/arch/x86_64/msr.rs
@@ -490,7 +490,7 @@ pub fn set_msrs(vcpu: &VcpuFd, msr_entries: &[kvm_msr_entry]) -> Result<(), MsrE
     vcpu.set_msrs(&msrs)
         .map_err(MsrError::SetMsrs)
         .and_then(|msrs_written| {
-            if msrs_written as u32 == msrs.as_fam_struct_ref().nmsrs {
+            if msrs_written == msrs.as_fam_struct_ref().nmsrs as usize {
                 Ok(())
             } else {
                 Err(MsrError::SetMsrsIncomplete)
